requirements.txt
pydantic
sqlalchemy
psycopg2-binary
python-dotenv